scala syntax
